pyproject.toml
[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "upwork_project1"
version = "0.1"
description = "A template project"
authors = ["Raoul van Hezel <raoulvanhezel@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
twelvedata = "^1.2.24"
alpaca-py = "^0.35.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
yfinance = "^0.2.52"
plotly = "^5.24.1"
ipywidgets = "^8.1.5"
holidays = "^0.65"
transformers = "^4.48.1"
torch = "^2.5.1"
websocket = "^0.2.1"
websocket-client = "^1.8.0"
